event page : modifications with dates
