+++ src/containers/EventForm.js
@@ -1,27 +1,25 @@
-import React from 'react';
-import { Button, Form, FormGroup, Label, Input, FormText } from 'reactstrap';
-import { Editor }       from 'react-draft-wysiwyg';
+import React      from 'react';
+import { Editor } from 'react-draft-wysiwyg';
 import 'react-draft-wysiwyg/dist/react-draft-wysiwyg.css';
 
-import FontAwesomeIcon   from '@fortawesome/react-fontawesome';
-import faSpinner         from '@fortawesome/fontawesome-free-solid/faSpinner';
+//Components
+import HRInfoSelect          from '../components/HRInfoSelect';
+import HRInfoLocations       from '../components/HRInfoLocations';
+import Address               from '../components/Address';
 
-import HRInfoSelect      from '../components/HRInfoSelect';
-import HRInfoLocations   from '../components/HRInfoLocations';
-import HRInfoAsyncSelect from '../components/HRInfoAsyncSelect';
-import RelatedContent    from '../components/RelatedContent';
-import HidContacts       from '../components/HidContacts';
-import Address           from '../components/Address';
-import EventDate         from '../components/EventDate';
-import LanguageSelect    from '../components/LanguageSelect';
-import StringSelect      from '../components/StringSelect';
+import HRInfoAsyncSelect     from '../components/HRInfoAsyncSelect';
+import HRInfoFiles           from '../components/HRInfoFiles';
+import RelatedContent        from '../components/RelatedContent';
+import LanguageSelect        from '../components/LanguageSelect';
+import EventCategorySelect   from '../components/EventCategorySelect';
+import EventDate             from '../components/EventDate';
 
 //Material plugin
 import FormHelperText   from '@material-ui/core/FormHelperText';
 import FormControl      from '@material-ui/core/FormControl';
 import FormLabel        from '@material-ui/core/FormLabel';
 import TextField        from '@material-ui/core/TextField';
-// import Button           from '@material-ui/core/Button';
+import Button           from '@material-ui/core/Button';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import Collapse         from '@material-ui/core/Collapse';
 import Card             from '@material-ui/core/Card';
@@ -34,200 +32,321 @@ import MomentUtils             from 'material-ui-pickers/utils/moment-utils';
 import MuiPickersUtilsProvider from 'material-ui-pickers/utils/MuiPickersUtilsProvider';
 import DatePicker              from 'material-ui-pickers/DatePicker';
 
+import './EventForm.css';
+
 class EventForm extends React.Component {
   constructor(props) {
     super(props);
-
     this.state = {
-      event_categories: [
-        { value: '82', label: 'Meetings'},
-        { value: '83', label: 'Trainings'},
-        { value: '84', label: 'Workshops'},
-        { value: '85', label: 'Conferences'}
-      ]
+      collapse     : false,
+      wasSubmitted : false,
     };
+
+    this.toggle    = this.toggle.bind(this);
+    this.hideAlert = this.hideAlert.bind(this);
+    this.submit    = this.submit.bind(this);
+  }
+
+  toggle() {
+    this.setState({ collapse: !this.state.collapse });
+  }
+
+  hideAlert() {
+    this.setState({ wasSubmitted: false });
+  }
+
+  submit() {
+    this.setState({ wasSubmitted: true });
   }
 
   render() {
-    const offices = this.props.doc.hasOperation ? (
-      <FormGroup>
-        <Label for="offices">Coordination hub(s)</Label>
-        <HRInfoSelect type="offices" spaces={this.props.doc.spaces} isMulti={true} onChange={(s) => this.props.handleSelectChange('offices', s)} value={this.props.doc.offices} />
-        <FormText color="muted">
-          Click on the field and select the coordination hub(s) the event refers to (if any).
-        </FormText>
-      </FormGroup>
-    ) : '';
+    const offices = this.props.doc.hasOperation
+    ? (
+      <FormControl fullWidth margin="normal">
+        <FormLabel>Coordination hub(s)</FormLabel>
+        <HRInfoSelect type     = "offices"
+                      spaces   = {this.props.doc.spaces}
+                      isMulti  = {true}
+                      onChange = {(s) => this.props.handleSelectChange('offices', s)}
+                      value    = {this.props.doc.offices}/>
+        <FormHelperText id = "offices-text">
+          Click on the field and select the coordination hub(s) the {this.props.typeLabel + ' '}
+          refers to (if any).
+        </FormHelperText>
+      </FormControl>
+    )
+    : '';
 
-    const disasters = this.props.doc.hasOperation ? (
-      <FormGroup>
-        <Label for="disasters">Disaster(s) / Emergency</Label>
-        <HRInfoSelect type="disasters" spaces={this.props.doc.spaces} isMulti={true} onChange={(s) => this.props.handleSelectChange('disasters', s)} value={this.props.doc.disasters} />
-        <FormText color="muted">
-          Click on the field and select the disaster(s) or emergency the event refers to. Each disaster/emergency is associated with a number, called GLIDE, which is a common standard used by a wide network of organizations See <a href="http://glidenumer.net/?ref=hrinfo">glidenumber.net</a>.
-        </FormText>
-      </FormGroup>
-    ) : '';
+    const disasters = this.props.doc.hasOperation
+    ? (
+      <FormControl fullWidth margin="normal">
+        <FormLabel>Disaster(s) / Emergency</FormLabel>
+        <HRInfoSelect type      = "disasters"
+                      spaces    = {this.props.doc.spaces}
+                      isMulti   = {true}
+                      onChange  = {(s) => this.props.handleSelectChange('disasters', s)}
+                      value     = {this.props.doc.disasters}/>
+        <FormHelperText id="disasters-text">
+            Click on the field and select the disaster(s) or emergency the {this.props.typeLabel + ' '}
+            refers to. Each disaster/emergency is associated with a number, called GLIDE, which is a common standard used by a wide network of organizations See
+            <a href="http://glidenumer.net/?ref=hrinfo"> glidenumber.net</a>.
+        </FormHelperText>
+      </FormControl>
+    )
+    : '';
 
-    const bundles = this.props.doc.hasOperation ? (
-      <FormGroup>
-        <Label for="bundles">Cluster(s)/Sector(s)</Label>
-        <HRInfoSelect type="bundles" spaces={this.props.doc.spaces} isMulti={true} onChange={(s) => this.props.handleSelectChange('bundles', s)} value={this.props.doc.bundles} />
-        <FormText color="muted">
-          Indicate the cluster(s)/sector(s) the event refers to.
-        </FormText>
-      </FormGroup>
-    ) : '';
+    const bundles = this.props.doc.hasOperation
+    ? (
+      <FormControl fullWidth margin="normal">
+        <FormLabel> Cluster(s)/Sector(s) </FormLabel>
+        <HRInfoSelect type     =  "bundles"
+                      spaces   =  {this.props.doc.spaces}
+                      isMulti  =  {true}
+                      onChange =  {(s) => this.props.handleSelectChange('bundles', s)}
+                      value    =  {this.props.doc.bundles}/>
+        <FormHelperText id="bundles-text">
+          Indicate the cluster(s)/sector(s) the {this.props.label + ' '}
+          refers to.
+        </FormHelperText>
+      </FormControl>
+    )
+    : '';
 
     return (
-      <Form onSubmit={this.props.handleSubmit} noValidate className={this.state.status === 'was-validated' ? 'was-validated bg-white my-3 p-3': 'bg-white my-3 p-3'}>
-        <FormGroup className="required">
-          <Label for="language">Language</Label>
-          <LanguageSelect value={this.props.doc.language} onChange={(s) => this.props.handleSelectChange('language', s)} className={this.props.isValid(this.props.doc.language) ? 'is-valid' : 'is-invalid'}/>
-          <div className="invalid-feedback">
-            Please select a language
-          </div>
-        </FormGroup>
+      <Grid container direction = "column" alignItems = "center">
+      <Typography color = "textSecondary" gutterBottom variant = "headline">Create {this.props.label}</Typography>
+      <Grid item>
+        <Grid container justify = "space-around">
+          <Grid item md ={6} xs ={11}>
 
-        <FormGroup className="required">
-          <Label for="spaces">Operation(s) / Webspace(s)</Label>
-          <HRInfoSelect type="spaces" isMulti={true} onChange={(s) => this.props.handleSelectChange('spaces', s)} value={this.props.doc.spaces} className={this.props.isValid(this.props.doc.spaces) ? 'is-valid' : 'is-invalid'} />
-          <FormText color="muted">
-            Click on the field and select where to publish the event (operation, regional site or thematic site).
-          </FormText>
-          <div className="invalid-feedback">
-            You must select an operation or a space
-          </div>
-        </FormGroup>
+          {/* Title */}
+            <FormControl required fullWidth margin = "normal">
+              <FormLabel focused error={this.props.status === 'was-validated' && !this.props.isValid(this.props.doc.label)}>Title</FormLabel>
+              <TextField type     = "text"
+                         name     = "label"
+                         id       = "label"
+                         value    = {this.props.doc.label}
+                         onChange = {this.props.handleInputChange}/>
+              <FormHelperText id = "label-text">
+                Type the original title of the {this.props.label + ' '}. Try not to use abbreviations. To see Standards and Naming Conventions click
+                <a href = "https://drive.google.com/open?id=1TxOek13c4uoYAQWqsYBhjppeYUwHZK7nhx5qgm1FALA"> here</a>.
+              </FormHelperText>
+            </FormControl>
 
-        <FormGroup className="required">
-          <Label for="label">Title</Label>
-          <Input type="text" name="label" id="label" placeholder="Enter the title of the event" required="required" value={this.props.doc.label} onChange={this.props.handleInputChange} />
-          <FormText color="muted">
-            Type the original title of the event. Try not to use abbreviations. To see Standards and Naming Conventions click <a href="https://drive.google.com/open?id=1TxOek13c4uoYAQWqsYBhjppeYUwHZK7nhx5qgm1FALA">here</a>.
-          </FormText>
-          <div className="invalid-feedback">
-            Please enter the event title
-          </div>
-        </FormGroup>
+         {/* Event Category */}
+            <FormControl required fullWidth margin="normal">
+              <FormLabel focused error = {this.props.status === 'was-validated' && !this.props.isValid(this.props.doc.category)}>Event Category</FormLabel>
+              <EventCategorySelect value     = {this.props.doc.category}
+                                   onChange  = {(s) => this.props.handleSelectChange('category', s)}
+                                   className = {this.props.isValid(this.props.doc.category) ? 'is-valid' : 'is-invalid'}/>
+              <FormHelperText id="language-text">
+                  From the list, select the kind of {this.props.label} you are creating. .
+              </FormHelperText>
+            </FormControl>
 
-        <FormGroup className="required">
-          <Label for="category">Event Category</Label>
-          <StringSelect id="category" name="category" options={this.state.event_categories} value={this.props.doc.category} onChange={(s) => this.props.handleSelectChange('category', s)} className={this.props.isValid(this.props.doc.category) ? 'is-valid' : 'is-invalid'}/>
-          <FormText color="muted">
-            From the list, select the kind of event you are creating.
-          </FormText>
-          <div className="invalid-feedback">
-            You must select an event category
-          </div>
-        </FormGroup>
+         {/* Event Description */}
+            <FormControl fullWidth margin = "normal">
+              <FormLabel>Event description</FormLabel>
+              <Card className = "card-container">
+                <Editor editorState         = {this.props.editorState}
+                        editorClassName     = "editor-content"
+                        toolbarClassName    = "editor-toolbar"
+                        onEditorStateChange = {this.props.onEditorStateChange}/>
+              </Card>
+              <FormHelperText id = "body-text">
+                Try to always include here the text (in full or part of it) of the {this.props.label + ' '}
+                (example: use the introduction or the executive summary). If no text is available add a description of the file(s) you are publishing.
+              </FormHelperText>
+            </FormControl>
 
-        <FormGroup className="required">
-          <Label for="date">Date(s)</Label>
-      /* */
-          <FormText color="muted">
-            Indicate the date(s) of the event.
-          </FormText>
-          <div className="invalid-feedback">
-            You must enter a date
-          </div>
-        </FormGroup>
+        {/* Address */}
+            <FormControl required fullWidth margin = "normal">
+              <FormLabel focused error ={this.props.status === 'was-validated' && !this.props.isValid(this.props.doc.files)}>Venue</FormLabel>
+              <Address onChange={(s) => this.props.handleSelectChange('files', s)} value={this.props.doc.files} />
+              <FormHelperText id = "files-text">
+                Indicate here where the event takes place.
+              </FormHelperText>
+            </FormControl>
+          </Grid>
 
-        <FormGroup>
-          <Label for="locations">Location</Label>
-          <HRInfoLocations onChange={(s) => this.props.handleSelectChange('location', s)} value={this.props.doc.location} />
-          <FormText color="muted">
-            Select from the menu the country(ies) the {this.state.typeLabel} is about and indicate more specific locations by selecting multiple layers (region, province, town).
-          </FormText>
-        </FormGroup>
+          <Grid item md={3} xs={11}>
+       {/* Language */}
+            <FormControl required fullWidth margin="normal">
+              <FormLabel focused error = {this.props.status === 'was-validated' && !this.props.isValid(this.props.doc.language)}>Language</FormLabel>
+              <LanguageSelect value     = {this.props.doc.language}
+                              onChange  = {(s) => this.props.handleSelectChange('language', s)}
+                              className = {this.props.isValid(this.props.doc.language) ? 'is-valid' : 'is-invalid'}/>
+              <FormHelperText id="language-text">
+                Select the language of the {this.props.label}.
+              </FormHelperText>
+            </FormControl>
 
-        <FormGroup>
-          <Label for="venue">Venue</Label>
-          <Address value={this.props.doc.address} onChange={(val) => this.props.handleSelectChange('address', val)} />
-          <FormText color="muted">
-            Indicate here where the event takes place.
-          </FormText>
-        </FormGroup>
+        {/* Operation(s) / Webspace(s) */}
+            <FormControl required fullWidth margin="normal">
+              <FormLabel focused error ={this.props.status === 'was-validated' && !this.props.isValid(this.props.doc.spaces)}>Operation(s) / Webspace(s)</FormLabel>
+              <HRInfoSelect type    = "spaces"
+                            isMulti = {true} onChange={(s) => this.props.handleSelectChange('spaces', s)}
+                            value   = {this.props.doc.spaces}/>
+              <FormHelperText>
+                Click on the field and select where to publish the {this.props.label + ' '}
+                (operation, regional site or thematic site).
+              </FormHelperText>
+            </FormControl>
 
-        <FormGroup>
-          <Label for="organizations">Organization(s)</Label>
-          <HRInfoAsyncSelect type="organizations" onChange={(s) => this.props.handleSelectChange('organizations', s)} value={this.props.doc.organizations} className={this.props.isValid(this.props.doc.organizations) ? 'is-valid' : 'is-invalid'} />
-          <FormText color="muted">
-            Type in and select the the event organizer(s).
-          </FormText>
-        </FormGroup>
+        {/* Dates */}
+            <FormControl required fullWidth margin="normal">
+              <FormLabel focused error={this.props.status === 'was-validated' && !this.props.isValid(this.props.doc.publication_date)}>Date</FormLabel>
+              <EventDate value    = {this.props.doc.date ? this.props.doc.date : ''}
+                         required
+                         onChange = {(val) => {console.log(val); this.props.handleSelectChange('date', val);}} />
+              <FormHelperText id = "publication_date-text">
+                Indicate the date of the {this.props.label + ' '}.
+              </FormHelperText>
+            </FormControl>
 
-        {bundles}
+        {/* Organizations */}
+            <FormControl required fullWidth margin="normal">
+              <FormLabel focused error={this.props.status === 'was-validated' && !this.props.isValid(this.props.doc.organizations)}>Organization(s)</FormLabel>
+              <HRInfoAsyncSelect type     = "organizations"
+                                 onChange = {(s) => this.props.handleSelectChange('organizations', s)}
+                                 value    = {this.props.doc.organizations}/>
+              <FormHelperText id="organizations-text">
+                Type in and select the source(s) of the {this.props.label}.
+              </FormHelperText>
+            </FormControl>
 
-        <FormGroup>
-          <Label for="body">Event description</Label>
-          <Editor
-            editorState={this.props.editorState}
-            wrapperClassName="demo-wrapper"
-            editorClassName="demo-editor"
-            onEditorStateChange={this.props.onEditorStateChange}
-          />
-          <FormText color="muted">
-            Provide here additional information regarding the event.
-          </FormText>
-        </FormGroup>
+        {/* Contacts */}
+            <FormControl required fullWidth margin="normal">
+              <FormLabel focused error={this.props.status === 'was-validated' && !this.props.isValid(this.props.doc.contacts)}>Contact(s)</FormLabel>
+              <HRInfoSelect type    = "spaces"
+                            isMulti = {true} onChange={(s) => this.props.handleSelectChange('contacts', s)}
+                            value   = {this.props.doc.contacts}/>
+              <FormHelperText>
+                Indicate the person(s) to contact for information regarding the {this.props.label}. To show up in the list, the person must have a HumanitarianID profile.
+              </FormHelperText>
+            </FormControl>
 
-        {offices}
-        {disasters}
-        <FormGroup>
-          <Label for="themes">Themes</Label>
-          <HRInfoSelect type="themes" isMulti={true} onChange={(s) => this.props.handleSelectChange('themes', s)} value={this.props.doc.themes} />
-          <FormText color="muted">
-            Click on the field and select all relevant themes. Choose only themes the event substantively refers to.
-          </FormText>
-        </FormGroup>
+        {/* Agenda(s) */}
+            <FormControl required fullWidth margin="normal">
+              <FormLabel focused error={this.props.status === 'was-validated' && !this.props.isValid(this.props.doc.agendas)}>Agenda(s)</FormLabel>
+              <HRInfoAsyncSelect type     = "documents"
+                                 onChange = {(s) => this.props.handleSelectChange('agendas', s)}
+                                 value    = {this.props.doc.agendas} />
+              <FormHelperText>
+                Add the agenda of the {this.props.label} as a document first, and then reference this document from here.
+              </FormHelperText>
+            </FormControl>
 
-        <FormGroup>
-          <Label for="contacts">Contact(s)</Label>
-          <HidContacts onChange={(s) => this.props.handleSelectChange('contacts', s)} value={this.props.doc.contacts} />
-          <FormText color="muted">
-            Indicate the person(s) to contact for information regarding the event. To show up in the list, the person must have a HumanitarianID profile.
-          </FormText>
-        </FormGroup>
+        {/* Meeting minute(s) */}
+            <FormControl required fullWidth margin="normal">
+              <FormLabel focused error={this.props.status === 'was-validated' && !this.props.isValid(this.props.doc.meeting_minutes)}>Meeting minute(s)</FormLabel>
+              <HRInfoAsyncSelect type     = "documents"
+                                 onChange = {(s) => this.props.handleSelectChange('meeting_minutes', s)}
+                                 value    = {this.props.doc.meeting_minutes} />
+              <FormHelperText>
+                Add the meeting minutes of the {this.props.label} as a document first, and then reference this document from here.
+              </FormHelperText>
+            </FormControl>
 
-        <FormGroup>
-          <Label for="agendas">Agenda(s)</Label>
-          <HRInfoAsyncSelect type="documents" onChange={(s) => this.props.handleSelectChange('agendas', s)} value={this.props.doc.agendas} />
-          <FormText color="muted">
-            Add the agenda of the event as a document first, and then reference this document from here.
-          </FormText>
-        </FormGroup>
+              {bundles}
+              {offices}
+              {disasters}
 
-        <FormGroup>
-          <Label for="meeting_minutes">Meeting minute(s)</Label>
-          <HRInfoAsyncSelect type="documents" onChange={(s) => this.props.handleSelectChange('meeting_minutes', s)} value={this.props.doc.meeting_minutes} />
-          <FormText color="muted">
-            Add the meeting minutes of the event as a document first, and then reference this document from here.
-          </FormText>
-        </FormGroup>
+        {/* Buttons show/hide more informations */}
+              <div className="more-info-button">
+                { !this.state.collapse &&
+                  <Button color="secondary" variant="contained" onClick={this.toggle}>
+                    <i className = "icon-plus" /> &nbsp; Add More Information
+                  </Button>
+                }
+                { this.state.collapse &&
+                  <Button color="secondary" variant="contained" onClick={this.toggle}>
+                    <i className = "icon-cancel" /> &nbsp; Hide Optional Information
+                  </Button>
+                }
+              </div>
 
-        <FormGroup>
-          <Label for="related_content">Related Content</Label>
-          <RelatedContent onChange={(s) => this.props.handleSelectChange('related_content', s)} value={this.props.doc.related_content} />
-          <FormText color="muted">
-            Add links to content that is related to the event you are creating by indicating the title of the content and its url. When using the Search function make sure to search by content title.
-          </FormText>
-        </FormGroup>
+              <Collapse in={this.state.collapse}>
+            {/* Locations */}
+                <FormControl fullWidth margin="normal">
+                  <FormLabel>Location(s)</FormLabel>
+                  <HRInfoLocations onChange = {(s) => this.props.handleSelectChange('locations', s)}
+                                   value    = {this.props.doc.locations}
+                                   isMulti  = "isMulti"
+                                   id       = "locations"/>
+                  <FormHelperText id="locations-text">
+                    Select from the menu the country(ies) the {this.props.label + ' '}
+                    is about and indicate more specific locations by selecting multiple layers (region, province, town).
+                  </FormHelperText>
+                </FormControl>
 
-        {this.state.status !== 'submitting' &&
+            {/* Themes */}
+                <FormControl fullWidth margin="normal">
+                  <FormLabel>Theme(s)</FormLabel>
+                  <HRInfoSelect type      = "themes"
+                                isMulti   = {true}
+                                onChange  = {(s) => this.props.handleSelectChange('themes', s)}
+                                value     = {this.props.doc.themes}
+                                id        = "themes"/>
+                  <FormHelperText id="themes-text">
+                    Click on the field and select all relevant themes. Choose only themes the {this.props.label + ' '}
+                    substantively refers to.
+                  </FormHelperText>
+                </FormControl>
+
+            {/* Related Content */}
+                <FormControl fullWidth margin="normal">
+                  <FormLabel>Related Content</FormLabel>
+                  <RelatedContent onChange  = {(s) => this.props.handleSelectChange('related_content', s)}
+                                  value     = {this.props.doc.related_content}
+                                  id        = "related_content"/>
+                  <FormHelperText id="related_content-text">
+                    Add links to content that is related to the {this.props.label + ' '}
+                    you are publishing (example: language versions of the same {this.props.label}, or the link of the event the meeting minutes refer to) by indicating the title of the content and its url.
+                  </FormHelperText>
+                </FormControl>
+              </Collapse>
+            </Grid>
+          </Grid>
+        </Grid>
+
+        <Grid item className="submission">
+        {
+          this.props.status !== 'submitting' &&
           <span>
-            <Button color="primary">Publish</Button> &nbsp;
-            <Button color="secondary" onClick={(evt) => this.props.handleSubmit(evt, 1)}>Save as Draft</Button> &nbsp;
+            <Button color="primary" variant="contained" onClick={(evt) => {this.props.handleSubmit(evt); this.submit()}}>Publish</Button>
+              &nbsp;
+            <Button color="secondary" variant="contained" onClick={(evt) => {this.props.handleSubmit(evt, 1); this.submit()}}>Save as Draft</Button>
+              &nbsp;
           </span>
         }
         {(this.props.status === 'submitting' || this.props.status === 'deleting') &&
-          <FontAwesomeIcon icon={faSpinner} pulse fixedWidth />
+          <CircularProgress />
         }
         {(this.props.match.params.id && this.props.status !== 'deleting') &&
-          <Button color="danger" onClick={this.props.handleDelete}>Delete</Button>
+          <span>
+            <Button color="secondary" variant="contained" onClick={this.props.handleDelete}>Delete</Button>
+          </span>
         }
-      </Form>
-    );
+        </Grid>
+        <Snackbar anchorOrigin={{
+            vertical   : 'bottom',
+            horizontal : 'left'
+          }}
+          open             = {this.props.status === 'was-validated' && this.state.wasSubmitted}
+          autoHideDuration = {6000}
+          onClose          = {this.hideAlert}
+          ContentProps     = {{
+            'aria-describedby' : 'message-id'
+          }}
+          message  = {<Typography id ="message-id" color="error">The form is incomplete and could not be submitted.</Typography>}
+          action   = {[
+            <Button key="undo" color="secondary" size="small" onClick={this.hideAlert}>
+            CLOSE
+            </Button>
+          ]}
+        />
+    </Grid>);
   }
 }
